[major] Added an async#each method.
This async#each method is different then most known iterators. It's loosely based
Zoidberg, a flow control library that I'm writing for Node.js. It features an
options argument that can be used to fine-tune the async methods. This can contain
the context of the callbacks as well as a execution timeout.